Fixed bug of sorted & duplicate data
Both dashboard and requests features were buggy in displaying sorted
data with time difference of 1ms. Currently, List is used instead of Set
(previous implementations), thus the result is expected to be prone of
duplicates, noting that the duplicates rest in the data, but is
different by 1ms in time.