Fixed basic authentication usages
Previous usages of basic authentication are not capable of returning
token and on every request, basic authentication must be sent. This
commit fixes those issues by issuing a token on login, and forces only 1
session per user, with need of the token existing on each request in
header with key X-Auth-Token.